requirements.txt
django==4.1.7
torchvision==0.9.0
pytorch==1.8.0
opencv-python==4.7.0.68
gunicorn
